Consolidate base viewmodel implementation
