--- app/src/main/java/com/topjohnwu/magisk/base/viewmodel/StatefulViewModel.kt
@@ -1,15 +1,0 @@
-package com.topjohnwu.magisk.base.viewmodel
-
-abstract class StatefulViewModel<State : Enum<*>>(
-    val defaultState: State
-) : ObservableViewModel() {
-
-    var state: State = defaultState
-        set(value) {
-            field = value
-            notifyStateChanged()
-        }
-
-    open fun notifyStateChanged() = Unit
-
-}
