+++ app/src/main/java/com/topjohnwu/magisk/ui/compat/CompatView.kt
@@ -2,8 +2,9 @@ package com.topjohnwu.magisk.ui.compat
 
 import android.view.View
 import androidx.core.graphics.Insets
+import com.topjohnwu.magisk.base.BaseViewModel
 
-internal interface CompatView<ViewModel : CompatViewModel> {
+internal interface CompatView<ViewModel : BaseViewModel> {
 
     val viewRoot: View
     val viewModel: ViewModel
